# [js/webgpu] Optimize InstanceNormalization