@dakenf 2023-09-11
My version was optimizing NHWC version and I did not touch the NCWH since it was not used right now, so i guess both changesets can be merged. I've used combined approach and calculated channel scale and channel shift using vec4/vec2 for different channels in a separate compute pass to speed-up applying these values https://github.com/microsoft/onnxruntime/blob/d45a96616da9843b037210f2d48d6b4e5bdae5c6/js/web/lib/wasm/jsep/webgpu/ops/instance-norm.ts