@qjia7 2023-09-12
@dakenf, I tried your method. The perf looks pretty well. But I find the results are different between our two methods. So we need some tests to guarantee the correctness (the existed model tests seems go to the cpu path not webgpu path due to opset is too low). For now, I will limit my optimization only on NCHW.