@dakenf 2023-09-12
You need to copy instance norm tests from web/test/data/node/opset8 to web/test/data/node/opset17 and then it will work fine. Not sure why in opset17 models are created with opset 6 (or some other low number)