@fs-eire 2023-09-12
The reason is because there was a commit in onnx: https://github.com/onnx/onnx/commit/acaa175e7dd48a4dfdc64f0fd3af7d2461f0b304.

This change makes each node test align with the operator's definition opset version. [InstanceNormalization](https://github.com/onnx/onnx/blob/main/docs/Operators.md#InstanceNormalization) has never changed since opset6, so that they make this node test of opset 6.

We don't want to modify onnx test data so I think we can create an op test for InstanceNormalization with the opset version we want.