@qjia7 2023-09-13
@fs-eire @guschmue Please take a look. Currently, only optimized the NCHW InstanceNormalization. Will leave NHWC to @dakenf. Thanks.